+++ edgy/cli/constants.py
@@ -1,8 +1,5 @@
 EDGY_DISCOVER_APP = "EDGY_DEFAULT_APP"
 DEFAULT_TEMPLATE_NAME = "default"
 APP_PARAMETER = "--app"
-HELP_PARAMETER = "--help"
-DISCOVERY_FILES = ["application.py", "app.py", "main.py"]
-DISCOVERY_FUNCTIONS = ["get_application", "get_app"]
-EXCLUDED_COMMANDS = ["list-templates", "inspectdb"]
-IGNORE_COMMANDS = ["list-templates"]
+DISCOVERY_PRELOADS = ["application", "app", "main", "asgi"]
+COMMANDS_WITHOUT_APP = ["list-templates", "inspectdb", "init"]