+++ docs/release-notes.md
@@ -6,6 +6,29 @@ hide:
 
 # Release Notes
 
+## 0.23
+
+### Added
+
+- Extension support via Monkay.
+
+### Changed
+
+- Rework edgy to use Monkay.
+- Rework Migrate system to simply use Monkay instance.
+- Replace `get_registry_copy` by `get_migration_prepared_registry`.
+- Breaking: migration configuration takes place in settings.
+- Breaking: EdgyExtra and Migrate are replaced by `edgy.Instance` but still available.
+- Breaking: `model_apps` is replaced by `preloads`.
+
+### Fixed
+
+- Migrations with ManyToMany fields are broken.
+
+### Unresolved
+
+- When copying a registry through models aren't copied properly.
+
 
 ## 0.22.0